requirements.txt
datasets==3.5.0
distilabel==1.5.3
huggingface_hub==0.30.1
langchain==0.3.25
langchain_chroma==0.2.3
langchain_ollama==0.3.2
langchain_text_splitters==0.3.8
openai==1.77.0
pydantic==2.11.4
python-dotenv==1.1.0
tqdm==4.67.1
transformers==4.51.0
trl==0.17.0
